RN-964: Break up submitSurvey into multiple files
