+++ packages/meditrak-app/app/assessment/actions/submitSurvey/index.js
@@ -1,0 +1,6 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+
+export { submitSurvey } from './submitSurvey';
